third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-pause/debugger-change-variable-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that modifying local variables works fine.

Set timer for test function.

Evaluated before modification:
localObject1.a = 310
localObject2 = 42

Evaluated after modification:
localObject1.a = -290
localObject2 = 123
